% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Crozier2005}
\alias{Crozier2005}
\title{Crozier and Tincani (2015)}
\format{
A data frame with 24 rows and 3 variables
}
\source{
Crozier, S., & Tincani, M. J. (2005). Using a modified social story to decrease disruptive behavior of a child with autism. 
Focus on Autism and Other Developmental Disabilities, 20(3), 150-157.
}
\description{
Data from an ABAB design that evaluated a social story intervention 
on the disruptive behavior of a preschool child with autism. The dependent variable, 
number of disruptions, was measured using frequency counting for a thirty minute observation
session. The variables are as follows:

\itemize{
\item \code{session}. Within-series session-number.
\item \code{phase}. Lable for each unique phase (e.g., A1 is the first baseline phase, B2 is the second treatment phase).
\item \code{score}. Value for the outcome.
}
}
\keyword{datasets}
