% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Dennis2021}
\alias{Dennis2021}
\title{Dennis & Whalon, (2021)}
\format{
A data frame with 90 rows and 5 variables
}
\source{
Dennis, L. R., & Whalon, K. J. (2021). Effects of teacher-versus application-delivered instruction on the expressive vocabulary of at-risk preschool children. Remedial and Special Education, 42(4), 195-206. doi:10.1177/0741932519900991
}
\description{
Data from a Repeated acquisition single-case experimental design evaluating vocabulary knowledge using an expressive definition task.
The variables are as follows:

\itemize{
  \item \code{Participant}. Participant identifier.
  \item \code{Condition}. Indicates condition, application (App) or teacher (TCH).
  \item \code{Observation}. Within-series session-number
  \item \code{Pre}. Pre-intervention score on the expressive definition task. 
  \item \code{Post}. Post-intervention score on the expressive definition task. 
}
}
\keyword{datasets}
