% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-measures.R
\name{LRR}
\alias{LRR}
\title{Log-response ratio}
\usage{
LRR(A_data, B_data, SE = TRUE, CI = TRUE, confidence = 0.95,
  bias_correct = TRUE)
}
\arguments{
\item{A_data}{vector of numeric data for A phase. Missing values are dropped.}

\item{B_data}{vector of numeric data for B phase. Missing values are dropped.}

\item{SE}{logical value indicating whether to report the standard error}

\item{CI}{logical value indicating whether to report a confidence interval}

\item{confidence}{confidence level for the reported interval estimate}

\item{bias_correct}{logical value indicating whether to use bias-correction.
Default is \code{TRUE}}
}
\value{
If \code{SE = FALSE} and \code{CI = FALSE}, a numeric value. 
  Otherwise, a list containing the estimate, standard error, and/or confidence 
  interval.
}
\description{
Calculates the log-response ratio effect size index, with or
  without bias correction (Pustejovsky, 2015)
}
\details{

}
\examples{

}
\references{
Pustejovsky, J. E. (2015). Measurement-comparable effect sizes 
  for single-case studies of free-operant behavior. \emph{Psychological
  Methods, 20}(3), 342--359.
  doi:\href{http://dx.doi.org/10.1037/met0000019}{10.1037/met0000019}
}

