% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCD-effect-sizes.R
\docType{data}
\name{McKissick}
\alias{McKissick}
\title{McKissick et al. (2010)}
\format{A data frame with 35 rows and 4 variables}
\source{
McKissick, C., Hawkins, R. O., Lentz, F. E., Hailley, J., & McGuire, S. (2010). 
Randomizing multiple contingency components to decrease disruptive behaviors and increase 
student engagement in an urban second-grade classroom. \emph{Psychology in the Schools, 47}(9), 
944–959. https://doi.org/10.1002/pits.20516
}
\description{
Data from a study by McKissick et al. (2010). The variables are as follows:

\itemize{
  \item \code{Case_Pseudonym}. Case Pseudonym provided by the authors.
  \item \code{Session_number}. Within-series session-number
  \item \code{Condition}. Describes whether the outcome is in the baseline (A) or treatment (B) phase.
  \item \code{Outcome}. Value for the outcome.
}
}
\keyword{datasets}
