% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Olszewski2017}
\alias{Olszewski2017}
\title{Olszewski, et al. (2017)}
\format{
A data frame with 72 rows and 4 variables
}
\source{
Olszewski, A., Soto, X., & Goldstein, H. (2017). Modeling alphabet
  skills as instructive feedback within a phonological awareness
  intervention. American Journal of Speech-Language Pathology, 26(3),
  769-790. doi:10.1044/2017_AJSLP-16-0042
}
\description{
Data from a multiple baseline design across behavior looking at
  four fluency measures (blends, segmenting, first part ID, first sound ID).
  The dependent variable is a score on a subtest, with a maximum score of 20.
  The variables are as follows:

\itemize{
  \item \code{behavior}. Outcome measure description (blends, segmenting, first part ID, or first sound ID)
  \item \code{session}. Within-series session-number
  \item \code{phase}. Label for each unique phase (A for baseline, B for treatment)
  \item \code{score}. Value for the outcome.
  \item \code{Include}. Value for inclusion in POGO calculation (0 if included in article, 1 if not included)
}
}
\keyword{datasets}
