% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Peters2020}
\alias{Peters2020}
\title{Peters-Sanders et al. (2020)}
\format{
A data frame with 153 rows and 4 variables
}
\source{
Peters-Sanders, L., Kelley, E., Haring, C., Madsen, K., Soto, X., Seven, Y., Hull, K., & Goldstein, H. (2020). Moving forward four words at a time: Effects of a supplemental preschool vocabulary. Language, Speech, and Hearing Services in Schools, 51, 165-175. https://doi.org/10.1044/2019_LSHSS-19-00029
}
\description{

}
\keyword{datasets}
