% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCD-effect-sizes.R
\name{SCD_effect_sizes}
\alias{SCD_effect_sizes}
\title{SCD effect size calculator}
\usage{
SCD_effect_sizes()
}
\description{
An interactive tool for calculating effect size indices for single-case designs.
}

