% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCD-effect-sizes.R
\docType{data}
\name{Schmidt2012}
\alias{Schmidt2012}
\title{Schmidt and Stichter (2012)}
\format{A data frame with 180 rows and 5 variables}
\source{
Schmidt, C., & Stichter, J. P. (2012). The use of peer-mediated interventions to 
promote the generalization of social competence for adolescents with high-functioning autism 
and Asperger's syndrome. \emph{Exceptionality}, 20(2), 94-113. doi:10.1080/09362835.2012
}
\description{
Data from an ABAB design conducted by Schmidt and 
Stichter (2012). All data were collected via continuous recording.
 The variables are as follows:

\itemize{
  \item \code{Case}. Participant identifier.
  \item \code{Behavior}. Behavior type (Conversation, Initiations, or Responses).
  \item \code{Trt} Treatment indicators.
  \item \code{Outcome}. Outcome scores.
  \item \code{Session_num}. Measurement occasion.
  \item \code{Session_length} Length of the observation session.
  \item \code{Procedure}  The metric in which the outcome measurement is expressed, all "other".
}
}
\keyword{datasets}
