% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradual-effects-model.R
\name{shine_gem_scd}
\alias{shine_gem_scd}
\title{Gradual Effects Model SCD Calculator}
\usage{
shine_gem_scd(browser = TRUE)
}
\arguments{
\item{browser}{logical value indicating whether to launch the app in the
system's default web-browser. Defaults to \code{TRUE}.}
}
\description{
Open an interactive tool for calculating the gradual effects model
for SCDs.
}
\examples{
\donttest{
shine_gem_scd()
}

}
