% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Spencer2012}
\alias{Spencer2012}
\title{Spencer et al. (2012)}
\format{
A data frame with 81 rows and 4 variables
}
\source{
Spencer, E., Goldstein, H., Sherman, A., Noe, S., Tabbah, R., Ziolkowski, R., & Schneider, N. (2012). 
Effects of an automated vocabulary and comprehension intervention: An early efficacy study. Journal of Early Intervention, 
34, 195-221. doi:10.1177/1053815112471990
}
\description{
Data from a repeated acquisition single-case experimental design 
evaluating vocabulary knowledge using an expressive definition task.
The variables are as follows:

\itemize{
\item \code{Observation}. Participant number.
\item \code{Book}. Book used for evaluation.
\item \code{Pre}. Pre-intervention score on the expressive definition task.
\item \code{Post}. Post-intervention score on the expressive definition task.
}
}
\keyword{datasets}
