% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Strasberger2014}
\alias{Strasberger2014}
\title{Strasberger & Ferreri (2014)}
\format{
A data frame with 47 rows and 8 variables
}
\source{
Strasberger, S. K., & Ferreri, S. J. (2014). The effects of peer
  assisted communication application training on the communicative and social
  behaviors of children with autism. \emph{Journal of Developmental and Physical
  Disabilities, 26}, 513-526. https://doi.org/10.1007/s10882-013-9358-9
}
\description{
Data from a multiple baseline design across participant that
  evaluated the effectiveness of an iPod-based speech generating device and a
  peer mediated intervention on improving sophisticated communication skills
  of children with autism. The variables are as follows:

\itemize{
  \item \code{StudyID}. Study identifier.
  \item \code{StudyID_CaseID}. Participant identifier.
  \item \code{Session_number}. Within-series session-number.
  \item \code{Condition}. Indicates condition, (A for baseline, B for intervention).
  \item \code{Outcome}. The number of independent 2-step mands.
  \item \code{Session_length}. Length of the session.
  \item \code{Procedure}. Scale of the outcome variable.
  \item \code{Goal_level}. Indicates the goal level for each series.
}
}
\keyword{datasets}
