% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-overlap-measures.R
\name{Tau}
\alias{Tau}
\title{Tau (non-overlap)}
\usage{
Tau(A_data, B_data, improvement = "increase", SE = TRUE,
  confidence = 0.05)
}
\arguments{
\item{A_data}{vector of numeric data for A phase}

\item{B_data}{vector of numeric data for B phase}

\item{improvement}{character string indicating direction of improvement.
Default is "increase"}

\item{SE}{logical value indicating whether to report the standard error of 
Tau}

\item{confidence}{confidence level for the reported interval estimate of Tau.
If \code{NULL} then no confidence interval is calculated.}
}
\value{
Numeric value. The range of Tau is [-1,1], with a null value of 0.
}
\description{
Calculates the Tau (non-overlap) index (Parker, Vannest, Davis, 
  & Sauber 2011).
}
\details{
Tau (non-overlap) is calculated as the Spearman rank-correlation 
  between the outcome observations and a binary variable indicating phase B. Tau
  is a linear re-scaling of NAP to the range [-1,1].
}
\examples{
A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
Tau(A_data = A, B_data = B)

}
\references{
Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. 
  (2011). Combining nonoverlap and trend for single-case research: Tau-U. 
  \emph{Behavior Therapy, 42}(2), 284--299. 
  doi:\href{http://dx.doi.org/10.1016/j.beth.2010.08.006}{10.1016/j.beth.2010.08.006}
}

