% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Thorne}
\alias{Thorne}
\title{Thorne and Kamps (2008)}
\format{A data frame with 776 rows and 5 variables}
\source{
Thorne, S., & Kamps, D. (2008). The effects of a group contingency intervention on academic engagement and problem 
behavior of at-risk students. \emph{Behavior Analysis in Practice, 1}(2), 12-18.
}
\description{
Data from an ABAB design conducted by Thorne and Kamps (2008). 
These data were used as an example in Swan and Pustejovsky (2017). Academic engagement 
was collected via continuous recording (marked as "other") and inappropriate 
verbalizations were collected via event counting (marked as "count").
The variables are as follows:

\itemize{
  \item \code{Measure}. Outcome measure description (academic engagement or inappropriate verbalizations).
  \item \code{Case}. Participant identifier.  
  \item \code{Session_number}. Measurement occasion.
  \item \code{Outcome}. Outcome scores
  \item \code{Trt}. Treatment indicators.
  \item \code{Session_length}. Length of the observation session.
  \item \code{Measure}. The metric in which the outcome measurement is expressed (count or other).
}
}
\references{
Swan, D. M., & Pustejovsky, J. E. (2017). 
A gradual effects model for single-case designs. http://doi.org/10.17605/OSF.IO/GAXRV
}
\keyword{datasets}
