% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Wright2012}
\alias{Wright2012}
\title{Wright & McCathren (2012)}
\format{A data frame with 98 rows and 5 variables}
\source{
Wright, L. A., & McCathren, R. B. (2012). 
Utilizing social stories to increase prosocial behavior and reduce 
problem behavior in young children with autism. 
\emph{Child Development Research}, 2012, 1-13. doi:10.1155/2012/357291
}
\description{
Data from a multiple baseline design conducted by Wright and
  McCathren (2012), which evaluated the effects of two types of social story
  interventions (a basic social story and a modified social story) on the
  behavior of four children with autism. Both dependent variables were
  measured using frequency counting for 10 minute observation sessions. The
  variables are as follows:

\itemize{
  \item \code{Participant} Participant identifier.
  \item \code{Session} Session number.
  \item \code{Condition} Phase of the design (baseline, intervention A, or intervention B.
  \item \code{Problem_behavior} Outcome scores.
}
}
\keyword{datasets}
