% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradual-effects-model.R
\name{gem_scd}
\alias{gem_scd}
\title{A function to estimate the gradual effects model for an SCD}
\usage{
gem_scd(Trt, outcome, m, fam)
}
\arguments{
\item{Trt}{A vector of treatment assignments where 0 = untreated and 1 =
treated.}

\item{outcome}{A vector of outcomes.}

\item{m}{The number of treatment occasions to estimate a treatment effect
for.}

\item{fam}{A description of the error distribution and link function to be
used in the model. This can be supplied in any way that `glm` will accept.
(See \code{\link[stats]{family}} for details of family functions.)}
}
\value{
`gem_scd` returns an object of class \code{glm}. In addition to the
  normal contents of a \code{glm} object, it also contains an estimate of
  `omega` and a variance covariance matrix called `varcov`.
}
\description{
A function to estimate the gradual effects model for an SCD
}
