% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Kelley2015}
\alias{Kelley2015}
\title{Kelley et al. (2015)}
\format{
A data frame with 54 rows and 5 variables
}
\source{
Kelley, E. S., Goldstein, H., Spencer, T., & Sherman, A. (2015). Effects of automated Tier 2 storybook intervention on vocabulary and comprehension learning in preschool children with limited oral language skills. Early Childhood Research Quarterly, 31, 47–61. https://doi.org/10.1016/j.ecresq.2014.12.004
}
\description{

}
\keyword{datasets}
