% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-measures.R
\name{LRR}
\alias{LRR}
\title{Log-response ratio}
\usage{
LRR(A_data, B_data, bias_correct = TRUE, confidence = 0.95)
}
\arguments{
\item{A_data}{vector of numeric data for A phase. Missing values are dropped.}

\item{B_data}{vector of numeric data for B phase. Missing values are dropped.}

\item{bias_correct}{logical value indicating whether to use bias-correction.
Default is \code{TRUE}}

\item{confidence}{confidence level for the reported interval estimate}
}
\value{
A list containing the estimate, standard error, and confidence 
  interval.
}
\description{
Calculates the log-response ratio effect size index, with or 
  without bias correction (Pustejovsky, 2015)
}
\details{
The response ratio parameter is the ratio of the mean level of the 
  outcome during phase B to the mean level of the outcome during phase A. The
  log response ratio is the natural logarithm of the response ratio. Without 
  bias correction, the log response ratio is estimated as the natural 
  logarithm of the phase B sample mean, minus the natural logarithm of the 
  phase A sample mean. A delta-method bias correction to the estimator is 
  used by default.
  
  The standard error of LRR is calculated based on a delta-method 
  approximation, allowing for the possibility of different degrees of 
  dispersion in each phase. The confidence interval for LRR is based on a
  large-sample (z) approximation.
}
\examples{
A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
LRR(A_data = A, B_data = B, bias_correct = FALSE)
LRR(A_data = A, B_data = B)

}
\references{
Pustejovsky, J. E. (2015). Measurement-comparable effect sizes 
  for single-case studies of free-operant behavior. \emph{Psychological 
  Methods, 20}(3), 342--359. 
  doi:\href{http://dx.doi.org/10.1037/met0000019}{10.1037/met0000019}
}
