% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-overlap-measures.R
\name{NAP}
\alias{NAP}
\title{Non-overlap of all pairs}
\usage{
NAP(A_data, B_data, improvement = "increase", SE = TRUE,
  confidence = 0.95)
}
\arguments{
\item{A_data}{vector of numeric data for A phase}

\item{B_data}{vector of numeric data for B phase}

\item{improvement}{character string indicating direction of improvement.
Default is "increase"}

\item{SE}{logical value indicating whether to report the standard error of 
NAP}

\item{confidence}{confidence level for the reported interval estimate of NAP.
If \code{NULL} then no confidence interval is calculated.}
}
\value{
Numeric value. The range of NAP is [0,1], with a null value of 0.5.
}
\description{
Calculates the non-overlap of all pairs index (Parker & Vannest,
  2009).
}
\details{
NAP is calculated as the proportion of all pairs of one observation
  from each phase in which the measurement from the B phase exceeds the
  measurement from the A phase, with pairs of data points that are exactly
  tied being given a weight of 0.5.
}
\examples{
A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
NAP(A_data = A, B_data = B)

}
\references{
Parker, R. I., & Vannest, K. J. (2009). An improved effect size 
  for single-case research: Nonoverlap of all pairs. \emph{Behavior Therapy,
  40}(4), 357--67.
  doi:\href{http://dx.doi.org/10.1016/j.beth.2008.10.006}{10.1016/j.beth.2008.10.006}
}

