% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-overlap-measures.R
\name{Tau}
\alias{Tau}
\title{Tau (non-overlap)}
\usage{
Tau(A_data, B_data, improvement = "increase", SE = TRUE, CI = TRUE,
  confidence = 0.95)
}
\arguments{
\item{A_data}{vector of numeric data for A phase. Missing values are dropped.}

\item{B_data}{vector of numeric data for B phase. Missing values are dropped.}

\item{improvement}{character string indicating direction of improvement.
Default is "increase"}

\item{SE}{logical value indicating whether to report the standard error}

\item{CI}{logical value indicating whether to report a confidence interval}

\item{confidence}{confidence level for the reported interval estimate}
}
\value{
A list containing the estimate, standard error, and/or confidence 
  interval.
}
\description{
Calculates the Tau (non-overlap) index (Parker, Vannest, Davis, 
  & Sauber 2011).
}
\details{
Tau (non-overlap) a linear re-scaling of \code{\link{NAP}} to the
  range [-1,1], with a null value of 0.
  
  Standard errors and confidence intervals for Tau are based on 
  transformations of the corresponding SEs and CIs for \code{\link{NAP}}
}
\examples{
A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
Tau(A_data = A, B_data = B)

}
\references{
Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. 
  (2011). Combining nonoverlap and trend for single-case research: Tau-U. 
  \emph{Behavior Therapy, 42}(2), 284--299. 
  doi:\href{http://dx.doi.org/10.1016/j.beth.2010.08.006}{10.1016/j.beth.2010.08.006}
}
