% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-overlap-measures.R
\name{Tau_BC}
\alias{Tau_BC}
\title{Tau-BC}
\usage{
Tau_BC(
  A_data,
  B_data,
  condition,
  outcome,
  baseline_phase = unique(condition)[1],
  improvement = "increase",
  SE = "unbiased",
  confidence = 0.95,
  pretest_trend = FALSE,
  report_correction = FALSE
)
}
\arguments{
\item{A_data}{vector of numeric data for A phase. Missing values are dropped.}

\item{B_data}{vector of numeric data for B phase. Missing values are dropped.}

\item{condition}{vector identifying the treatment condition for each
observation in the series.}

\item{outcome}{vector of outcome data for the entire series.}

\item{baseline_phase}{character string specifying which value of
\code{condition} corresponds to the baseline phase. Defaults to first
observed value of \code{condition}.}

\item{improvement}{character string indicating direction of improvement.
Default is "increase".}

\item{SE}{character value indicating which formula to use for calculating the
standard error of Tau-BC, with possible values \code{"unbiased"} for the
exactly unbiased estimator, \code{"Hanley"} for the Hanley-McNeil
estimator, \code{"null"} for the (known) variance under the null hypothesis
of no effect, or \code{"none"} to not calculate a standard error. Defaults
to "unbiased". Note that the "unbiased" standard error is unbiased for
\code{\link{Tau}}, but not necessarily unbiased for \code{\link{Tau_BC}}.
None of the standard error formulas account for the additional uncertainty
due to use of the baseline trend correction.}

\item{confidence}{confidence level for the reported interval estimate. Set to
\code{NULL} to omit confidence interval calculations.}

\item{pretest_trend}{significance level for the initial baseline trend test.
The raw data are corrected and \code{\link{Tau_BC}} is calculated only if
the baseline trend is statistically significant. Otherwise,
\code{\link{Tau_BC}} is equal to \code{\link{Tau}}. Default is
\code{FALSE}, which always adjusts for the baseline trend.}

\item{report_correction}{logical value indicating whether to report the
baseline corrected slope and intercept values. Default is \code{FALSE}.}
}
\value{
A list containing the estimate, standard error, and/or confidence
  interval.
}
\description{
Calculates the baseline-corrected Tau index (Tarlow 2017).
}
\details{
Tau-BC is an elaboration of the \code{\link{Tau}} that includes a
  correction for baseline trend. The calculation of Tau-BC involves two or
  three steps, depending on the \code{pretest_trend} argument.

  If \code{pretest_trend = FALSE} (the default), the first step involves
  adjusting the outcomes for baseline trend estimated using Theil-Sen
  regression. In the second step, the residuals from Theil-Sen regression are
  used to calculate the \code{\link{Tau}} (non-overlap) index.

  If \code{pretest_trend = a}, for some significance level \code{a} between 0
  and 1, then the firs step involves a significance test for the slope of the
  baseline trend based on Kendall's rank correlation. If the slope is not
  significantly different from zero, then no baseline trend adjustment is
  made and Tau-BC is set equal to \code{\link{Tau}}. If the slope is
  significantly different from zero, then in the second step, the outcomes
  are adjusted for baseline trend using Theil-Sen regression and, in the
  third step, the residuals from Theil-Sen regression are used to calculate
  the \code{\link{Tau}} (non-overlap) index.

  Note that the standard error formulas are based on the standard errors for
  \code{\link{Tau}} (non-overlap) and they do not account for the additional
  uncertainty due to use of the baseline trend correction (nor to the
  pre-test for statistical significance of baseline trend, if used).
}
\examples{
A <- c(20, 20, 26, 25, 22, 23)
B <- c(28, 25, 24, 27, 30, 30, 29)
Tau_BC(A_data = A, B_data = B)

}
\references{
Tarlow, K. R. (2017). An improved rank correlation effect size
  statistic for single-case designs: Baseline corrected Tau. \emph{Behavior
  modification, 41}(4), 427-467. doi:\doi{10.1177/0145445516676750}
}
